Move getLocalizedStringEN to presentation-backend
