+++ presentation/backend/src/presentation-backend/Utils.ts
@@ -11,6 +11,21 @@ import { Element, IModelDb } from "@itwin/core-backend";
 import { DbResult, Id64String } from "@itwin/core-bentley";
 import { Diagnostics, DiagnosticsOptions, InstanceKey } from "@itwin/presentation-common";
 
+const ecPresentation = require("@itwin/presentation-common/assets/locales/en/ECPresentation.json"); // eslint-disable-line @typescript-eslint/no-var-requires
+const rulesEngine = require("@itwin/presentation-common/assets/locales/en/RulesEngine.json"); // eslint-disable-line @typescript-eslint/no-var-requires
+
+/** @internal */
+export function getLocalizedStringEN(key: string) {
+  const [namespace, identifier] = key.split(":", 2);
+  if (namespace === "ECPresentation") {
+    return ecPresentation[identifier];
+  }
+  if (namespace === "RulesEngine") {
+    return rulesEngine[identifier];
+  }
+  return key;
+}
+
 /** @internal */
 export function getElementKey(imodel: IModelDb, id: Id64String): InstanceKey | undefined {
   let key: InstanceKey | undefined;
