+++ presentation/backend/src/test/Utils.test.ts
@@ -7,7 +7,7 @@ import * as moq from "typemoq";
 import { DbResult } from "@itwin/core-bentley";
 import { ECSqlStatement, ECSqlValue, IModelDb } from "@itwin/core-backend";
 import { createRandomId } from "@itwin/presentation-common/lib/cjs/test";
-import { getElementKey, normalizeVersion } from "../presentation-backend/Utils";
+import { getElementKey, getLocalizedStringEN, normalizeVersion } from "../presentation-backend/Utils";
 
 describe("getElementKey", () => {
 
@@ -62,6 +62,31 @@ describe("getNormalizedVersion", () => {
 
   it("returns `0.0.0` on invalid version string", () => {
     expect(normalizeVersion("invalid")).to.eq("0.0.0");
+  });
+
+});
+
+describe("getLocalizedStringEN", () => {
+
+  it("translates from RulesEngine", () => {
+    expect(getLocalizedStringEN("RulesEngine:LABEL_General_NotSpecified")).to.be.eq("Not specified");
+    expect(getLocalizedStringEN("RulesEngine:LABEL_General_Other")).to.be.eq("Other");
+    expect(getLocalizedStringEN("RulesEngine:LABEL_General_Varies")).to.be.eq("Varies");
+    expect(getLocalizedStringEN("RulesEngine:LABEL_General_MultipleInstances")).to.be.eq("Multiple items");
+  });
+
+  it("translates from ECPresentation", () => {
+    expect(getLocalizedStringEN("ECPresentation:LABEL_General_DisplayLabel")).to.be.eq("Label");
+    expect(getLocalizedStringEN("ECPresentation:LABEL_Category_General")).to.be.eq("General");
+    expect(getLocalizedStringEN("ECPresentation:LABEL_Category_Favorite")).to.be.eq("Favorite");
+    expect(getLocalizedStringEN("ECPresentation:FIELD_LABEL_DisplayLabel")).to.be.eq("Label");
+    expect(getLocalizedStringEN("ECPresentation:CATEGORY_LABEL_SelectedItems")).to.be.eq("Selected Item(s)");
+    expect(getLocalizedStringEN("ECPresentation:CATEGORY_DESCRIPTION_SelectedItems")).to.be.eq("Contains properties of selected item(s)");
+    expect(getLocalizedStringEN("ECPresentation:ERROR_General_Unknown")).to.be.eq("Unknown error");
+  });
+
+  it("does not translate if key not found", () => {
+    expect(getLocalizedStringEN("wrong:Label")).to.be.eq("wrong:Label");
   });
 
 });
