+++ presentation/common/src/test/LocalizationHelper.test.ts
@@ -5,7 +5,7 @@
 
 import { expect } from "chai";
 import { createRandomECInstancesNode, createRandomLabelCompositeValue, createRandomLabelDefinition, createTestCategoryDescription, createTestContentDescriptor, createTestContentItem, createTestECInstanceKey, createTestSimpleContentField } from "./_helpers";
-import { Content, getLocalizedStringEN, Item, LabelDefinition, LocalizationHelper } from "../presentation-common";
+import { Content, Item, LabelDefinition, LocalizationHelper } from "../presentation-common";
 
 function getTestLocalizedString(key: string) {
   if (key.includes(":"))
@@ -203,31 +203,6 @@ describe("LocalizationHelper", () => {
       });
     });
 
-  });
-
-});
-
-describe("getLocalizedStringEN", () => {
-
-  it("translates from RulesEngine", () => {
-    expect(getLocalizedStringEN("RulesEngine:LABEL_General_NotSpecified")).to.be.eq("Not specified");
-    expect(getLocalizedStringEN("RulesEngine:LABEL_General_Other")).to.be.eq("Other");
-    expect(getLocalizedStringEN("RulesEngine:LABEL_General_Varies")).to.be.eq("Varies");
-    expect(getLocalizedStringEN("RulesEngine:LABEL_General_MultipleInstances")).to.be.eq("Multiple items");
-  });
-
-  it("translates from ECPresentation", () => {
-    expect(getLocalizedStringEN("ECPresentation:LABEL_General_DisplayLabel")).to.be.eq("Label");
-    expect(getLocalizedStringEN("ECPresentation:LABEL_Category_General")).to.be.eq("General");
-    expect(getLocalizedStringEN("ECPresentation:LABEL_Category_Favorite")).to.be.eq("Favorite");
-    expect(getLocalizedStringEN("ECPresentation:FIELD_LABEL_DisplayLabel")).to.be.eq("Label");
-    expect(getLocalizedStringEN("ECPresentation:CATEGORY_LABEL_SelectedItems")).to.be.eq("Selected Item(s)");
-    expect(getLocalizedStringEN("ECPresentation:CATEGORY_DESCRIPTION_SelectedItems")).to.be.eq("Contains properties of selected item(s)");
-    expect(getLocalizedStringEN("ECPresentation:ERROR_General_Unknown")).to.be.eq("Unknown error");
-  });
-
-  it("does not translate if key not found", () => {
-    expect(getLocalizedStringEN("wrong:Label")).to.be.eq("wrong:Label");
   });
 
 });
